Added text size get intent method

+++ app/src/main/java/edu/temple/activities/TextSizeActivity.kt
@@ -31,7 +31,7 @@ class TextSizeActivity : AppCompatActivity() {
                     Intent(
                         this@TextSizeActivity, DisplayActivity::class.java
                     ).apply{
-                        putExtra(FONT_SIZE_KEY, it.toFloat())
+                        putExtra(FONT_SIZE_KEY, it)
                     }
                 )
